Split model page specs and copy helpers
