+++ frontend/app/(localized)/[locale]/(marketing)/models/[slug]/_lib/model-page-spec-status.ts
@@ -1,0 +1,91 @@
+import type { AppLocale } from '@/i18n/locales';
+import { SPEC_STATUS_LABELS } from './model-page-specs-constants';
+
+export function isPending(value: string) {
+  return value.trim().toLowerCase() === 'data pending';
+}
+
+export function isUnsupported(value: string) {
+  const normalized = value.trim().toLowerCase();
+  return normalized === 'not supported' || normalized === 'unsupported';
+}
+
+export function isSupported(value: string) {
+  return value.trim().toLowerCase() === 'supported';
+}
+
+export function resolveSpecStatusLabels(locale: AppLocale) {
+  return SPEC_STATUS_LABELS[locale] ?? SPEC_STATUS_LABELS.en;
+}
+
+export function localizeSpecStatus(value: string, locale: AppLocale): string {
+  const labels = resolveSpecStatusLabels(locale);
+  const normalized = value.trim();
+  const lower = normalized.toLowerCase();
+  if (isSupported(normalized)) return labels.supported;
+  if (isUnsupported(normalized)) return labels.notSupported;
+  if (isPending(normalized)) return labels.pending;
+  if (lower.startsWith('supported (') && normalized.endsWith(')')) {
+    const detail = normalized.slice(normalized.indexOf('(') + 1, -1);
+    return `${labels.supported} (${localizeSpecStatus(detail, locale)})`;
+  }
+  if (lower.startsWith('not supported (') && normalized.endsWith(')')) {
+    const detail = normalized.slice(normalized.indexOf('(') + 1, -1);
+    return `${labels.notSupported} (${localizeSpecStatus(detail, locale)})`;
+  }
+  if (lower === 'prompt-based only') {
+    return locale === 'fr' ? 'Via prompt uniquement' : locale === 'es' ? 'Solo mediante prompt' : normalized;
+  }
+  if (lower === 'single start image') {
+    return locale === 'fr' ? 'une seule image de départ' : locale === 'es' ? 'una sola imagen inicial' : normalized;
+  }
+  if (lower === 'multi reference stills') {
+    return locale === 'fr'
+      ? 'plusieurs stills de référence'
+      : locale === 'es'
+        ? 'varios stills de referencia'
+        : normalized;
+  }
+  if (lower === 'source clip for extend / retake') {
+    return locale === 'fr'
+      ? 'clip source pour extension / retake'
+      : locale === 'es'
+        ? 'clip fuente para extensión / retake'
+        : normalized;
+  }
+  if (lower === 'source clip for modify / reframe') {
+    return locale === 'fr'
+      ? 'clip source pour modify / reframe'
+      : locale === 'es'
+        ? 'clip fuente para modify / reframe'
+        : normalized;
+  }
+  if (lower === 'start + end image in i2v') {
+    return locale === 'fr'
+      ? 'image de départ + image de fin en image → vidéo'
+      : locale === 'es'
+        ? 'imagen inicial + imagen final en imagen → video'
+        : normalized;
+  }
+  if (lower === 'reframe workflow') {
+    return locale === 'fr' ? 'workflow reframe' : locale === 'es' ? 'workflow reframe' : normalized;
+  }
+  if (lower === 'modify / reframe workflows') {
+    return locale === 'fr'
+      ? 'workflows modify / reframe'
+      : locale === 'es'
+        ? 'workflows modify / reframe'
+        : normalized;
+  }
+  if (lower === 'extend / retake workflows') {
+    return locale === 'fr'
+      ? 'workflows extension / retake'
+      : locale === 'es'
+        ? 'workflows de extensión / retake'
+        : normalized;
+  }
+  if (lower === 'no (maxvideoai)') {
+    return locale === 'fr' ? 'Non (MaxVideoAI)' : locale === 'es' ? 'No (MaxVideoAI)' : normalized;
+  }
+  return value;
+}
